% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplefunction.R
\name{samplefunction}
\alias{samplefunction}
\title{Sample Function Title}
\usage{
samplefunction(x, y = 10, z = c("yellow", "red", "green"))
}
\arguments{
\item{x}{This is a required argument and has no default value.}

\item{y}{Has a default value of 10.}

\item{z}{Although this looks like it is a default string, the function \code{match.arg()} is
used to set it in the function. With \code{match.arg()}, the string specifies what values are allowed and it will throw an error if not one of those. The default is the first value.}
}
\value{
A string with the values.
}
\description{
This part is the description. It can be as long as you want but usually is one paragraph.
}
\examples{
samplefunction(1)
}
