% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleplot.R
\name{sampleplot}
\alias{sampleplot}
\title{Sample plot}
\usage{
sampleplot(x, y)
}
\arguments{
\item{x}{The x variables.}

\item{y}{The y response values.}
}
\value{
the coefficient of the linear regression
}
\description{
This shows how you need to use \code{::} with all your functions that are not in base R.
}
\examples{
x <- 1:10
y <- 10+3*x + stats::rnorm(10, 0, 10)
sampleplot(x, y)
}
